chore: Moves Dropdown to the components folder
