+++ superset-frontend/src/common/components/common.stories.tsx
@@ -23,10 +23,10 @@ import Button from 'src/components/Button';
 import { CronPicker, CronError } from 'src/components/CronPicker';
 import Modal from 'src/components/Modal';
 import InfoTooltip from 'src/components/InfoTooltip';
+import { Dropdown } from 'src/components/Dropdown';
 import Tabs, { EditableTabs } from './Tabs';
 import { Tooltip as AntdTooltip } from './Tooltip';
 import { Menu, Input, Divider } from '.';
-import { Dropdown } from './Dropdown';
 
 export default {
   title: 'Common components',
